windowManager: make DesktopOverlay track transient windows too
The overlay cover we use to trap clicks outside of the side component,
when it's triggered from the application grid, doesn't extend to cover
transient children of the side component - e.g. a file chooser.
We now extend the DesktopOverlay mechanism to account for that, and we
also move it to a separate class, to keep WindowManager clean.

[endlessm/eos-shell#2786]