iconGridLayout: animate icon undeletion
Instead of keeping the whole previous layout
so it can be restored when the user uses "Undo", just keep
the position of the removed item and use it to reposition it
in its previous place.

appDisplay: fix _findIconChanges() so it lists removed items.

iconGrid: split out the code calculating a child's position
from _allocate(). When new icons are going to be inserted in
the grid, we use this _childAllocation() method to calculate
the target positions of existing icons that need to be
animated as they move into new slots.

[endlessm/eos-shell#2086]