Record metric on desktop search
When the user does a Google search or a local search from the desktop,
record a metric.

The Google search is executed only once when the user presses Enter, so
the metric is recorded at the same time that the search is executed.

The local search is live, and a search is queued for each keystroke in
the entry, so for the local search we only record a metric 3 seconds
after the user has stopped typing. This is in order not to waste space
and bandwidth with metrics for searching for q, qu, que, quer, query.

[endlessm/eos-sdk#868]