appIconBar: Use a separate PopupMenuManager for the two menus
PopupMenuManager is woefully inept at having two menus with the same
source actor. When the source actor gets hovered over or key focus,
we'll try to pop up the menu associated with the source actor. When
there's two of them, they'll keep signalling and looping around,
causing the other popup to show up on every frame.

Instead of fixing PopupMenuManager to deal with this weird edge case,
just use a separate PopupMenuManager, which means that only one at a
time gets signals, and the two menus can't fight.

[endlessm/eos-shell#812]