Revert "appDisplay: sanitize app IDs before passing them to layout"
This reverts commit b0596a1ad81711feadba6033faca85a06acc8a38.

This commit is wholly wrong.

For core apps like gnome-terminal, we ship both the upstream
gnome-terminal.desktop and eos-app-gnome-terminal.desktop.

The .desktop file that appears in the app store and will appear on the
desktop will always be the prefixed one, eos-app-*. The upstream
.desktop file is unused, and so we should always prefer the eos-app-*
branded one.

As such, stripping eos-app-* from .desktop results is wrong, as the
unprefixed variant can't ever possibly be in the icon grid layout.

Restore the original code to make core apps work again.