Don't show splash screen unless explicitly enabled
Only those apps with desktop files that specify to launch
maximized will show a splash screen.

This fixes an issue with launching gnome-control-center
from the desktop search, where we get an undesired splash screen.
Even worse, if gnome-control-center was already open,
the splash screen was stuck open in the foreground.

Originally, when we added this feature, the intention was
to always show a splash screen unless explicitly blacklisted.
In reality, though, we now explicitly generate .desktop files
for everything that is launchable from the desktop, so we
have the desktop key for everything we want to launch maximized
with the splash screen.

[endlessm/eos-shell#3947]