workspaceMonitor: cache fullscreen state
We want to be robust to spurious emissions of the
"in-fullscreen-changed" signal on MetaScreen, so cache the fullscreen
state in WorkspaceMonitor.

This fixes a case where we would wrongly call updateOverview() with
zero visibleWindows due to a spurious emission of the MetaScreen signal,
which would result being left back at the desktop grid after launching
an application right after login.

[endlessm/eos-shell#3736]