appStore: add a hack to avoid showing overview when an app launches
We currently hide the app store window when launching an app. In case
the application is not running, we get a splash screen, and then the
application is shown, and everything is fine.

When the application is already running though, we currently just
activate its window and bring it to the front. But when the app store is
then hidden, if it was launched from the desktop, the code in
WindowManager will show back the desktop, as it has no knowledge that in
the meantime another application was launched. Add that knowledge to the
code, even if it doesn't really look pretty.

[endlessm/eos-shell#4034]