Remove auto connection button for modems
The code for this feature ("Auto broadband"
or "Auto dial-up") was not implemented,
such that clicking the button led to a missing
pure virtual method and thus had no action.
If implemented, the action taken would be the
same as sliding the switch from off to on,
so there is no need for this button.

[endlessm/eos-shell#4279]