Added evaluation of initial desktop area hover state in overview
Previous code did not evaluate correctly if the pointer was over the
desktop area on initial show of the overview but could handle
transition states. This change fixes that but we still need handling of
window -> overview transition and maybe tween fix on animating the hover
state of the mentioned area.