Added singal handlers for desktop area hover when showing app windows
This adds signal handlers to apply the hover effect on the desktop area
if it's hovered while a non-maximized app window is displayed. While the
code should work as expected, mutter handling of events in this case is
not working properly but once that is fixed, things should work fine.