runDialog: don't complete results from the home directory
Currently the run dialog expands results from the home directory,
regardless of whether they're executables in paths or just files in the
directory.
Since we already have a search provider that takes care of that, do not
pollute the completion with things that might be in the home directory.
This won't remove the ability to open a file in the home directory, or
elsewhere, with its default application, by specifying a real path
inside the dialog prompt.

[endlessm/eos-shell#4524]