overview: super key toggles desktop search active
Now in the Overview class we have two toggle functions, one for
search and one for the window picker.

The keybinding for panel-main-menu continues to be bound to the
window picker (by default Super+S), and the super key will now
toggle the desktop visible with the search bar in focus.
[endlessm/eos-shell#3199]