background: only disconnect file monitor when there are no clients
Currently we immediately disconnect file monitor events after we receive
the first 'changed' signal. This doesn't really work, since file
monitors can be shared between different payloads in the cache, as long
as they all point to the same file.

Extend the existing logic in removeImageContent() that deletes the file
monitor when there are no other clients to also disconnect events when
that condition is verified.

[endlessm/eos-shell#4586]