background: rework method to set a new background image
Currently BackgroundManager has two fields for Background objects:
this.background and this._newBackground. The former represents the
background as it's current, where the latter represents the next
background we're transitioning to.

There's a few problems with this setup:
- in _updateBackground() we connect to the "loaded" signal after
  creating the new background, but we never check whether the object is
  loaded already. This can happen if we're loading from cache.
- in case another change is received while the new background is being
  loaded, we try to catch that by dropping the previous change on the
  floor and destroying the previous actor. This interacts badly with
  the tween animation that we're doing.
- the Tweener function will try to set this.background to the
  newBackground object, but it does not take into account the case
  above, so it will fail when another change happens mid-flight.

Avoid this complications entirely by making sure that this.background is
always the most up-to-date background object; loaded vs non-loaded is
taken care of inside the Background class already.
Also change the tween out function so that it always tweens the
previously displayed background.

This effectively gets rid of this._newBackground and fixes races
conditions observed while quickly changing the background GSettings
while an animation is still in progress.

[endlessm/eos-shell#4586]