overview: Transition instantly to active windows from search
When transitioning from the icon grid to active windows we always
transitioned instantly. The search transition was oddly inconsistent
and faded the desktop to a completely blank state which was even more
odd.

Now we do the same instant transition from the desktop regardless of
whether we are on the icon grid or search page.
[endlessm/eos-shell#4606]