Remove all trace of the 'separate-power-off-log-out' GSettings option
This option has been introduced to fix endlessm/shell#1010, when the
decision to hide the Log Out button was made, as a way to allow
restoring the original behaviour. However, now that we moved to round
buttons it does not make sense at all, at least not without further
input from design or without considering other options such as the
user switcher from GNOME 3.10.

[endlessm/eos-shell#4170]