viewSelector: move search bar above icon grid
The search bar will sit half way between the top of the icon grid
and top of the screen. When the search is activated the entry will
animate toward the top of the screen, while the icon grid fades out
and results fade in.

In the process simplified the views clone a bit so it never needs
a ViewsDisplayContainer. It just uses the ViewsDisplayLayout
directly.
[endlessm/eos-shell#4606]