viewSelector: freeze search results actor while animating
Updating the search UI is expensive routine, especially with the
text shadows we have on most of our results. There's no way to
get consistent frame rates on the search animation while drawing
updated results. And its a bit too much changing on screen from
a user perspective. So we force the search to stop updating its
results while animating.

Also, we never show the status bin while animating, because that
looks ugly.
[endlessm/eos-shell#4606]