search: don't clear results after empty search
We now need to fade the search out after it get unactivated, so
we can't clear the search results as soon as we receive an
empty set of terms.

On the first search following an empty search, we clear the results
and immediately launch a new search with no timeout. This way we can
show some results to the user as quick as possible.
[endlessm/eos-shell#4606]