background: check for other users after removing our user
Code currently checks whether there are other users of a file to
determine when to stop monitoring it in the cache.
By the time the check runs though, we have not yet removed our own user,
so the check will always pass. This will cause us to leave lingering
file monitors when the background changes.

Move the check after we have removed our own user, so that we'll
correctly exclude ourselves from the count.

[endlessm/eos-shell#4586]