search: use simplified search icon for app/provider results
We had two separate app icons on our search results:
 - provider icons, which launched a search within an app
 - app icons, from our application search, which just lunch the app

The app icons were a bit broken, they used the exact same icons as
on our icon grid. Which meant they had two separate hover state
(the search one and the hover state for the desktop icons), as well
as two clicked handlers which did the same thing.

Provider icons lacked labels, which we want in our new search theme.

This commit consolidates the two to use a simplified SearchIcon class,
that can be themed separately from our desktop icons.

This kills the ability to drag and drop to the dash, but we have plans
to remove the dash anyway, so didn't worry about preserving that.
[endlessm/eos-shell#4719]