search: don't give single provider results extra space
We used to pack individual results so that they would expand to
grab extra space in the list provider. This meant that providers
with only a single result would have a large and weird looking
allocation for that result.

Instead we will just center results in the list provider without
expanding, so that all results should have the same allocation
[endlessm/eos-shell#4719]