search: When animating, update results before activating one
We freeze the results while animating the search upwards to avoid
bombarding the user with changing elements and to keep frame rates
smooth.

If our results are activated while the animation is happening, we
need to unfreeze the results first. Otherwise we will launch the
default result for the query entered when the animation started,
instead of the current query.
[endlessm/eos-shell#4767]