viewSelector: move icon grid placement down slightly
We want the icons inside the icon grid to appear perfectly centered
in the desktop area, so we add some padding to the top of the icon
grid when possible, to offset the presence of the icon labels at
the bottom of the grid.

With our current sizing model, the ViewsDisplayLayout seemed the
only place to handle this correctly.
[endlessm/eos-shell#4874]