shellEntry: backspace event synthesized when not focused
When we receive a keypress event on the stage that should trigger
a search, we synthesize that event of the search entry itself.

This was working in all cases except backspace event, which appeared
to be captured by another actor. We shouldn't synthesize these key
events in the capture phase, as we already we want the entry handling
those events.
[endlessm/eos-shell#4898]