Define SHELL_APP_IS_OPEN_EVENT locally.
EMTR_EVENT_SHELL_APP_IS_OPEN is deprecated because event IDs should be
defined in the file in which they are used. Otherwise we run into
problems when changing the wire protocol of those event IDs since the
event ID needs to change in lock step with the auxiliary payload.

[endlessm/eos-sdk#3029]