misc: Add missing check on reply from CanReboot/CanPowerOff D-Bus methods
According to systemd logind's documentation, these two methods "return
'na' in case "the operation is not available because hardware, kernel
or drivers do not support it", while "'no' is returned if the operation
is available but the user is not allowed to execute it".

See http://www.freedesktop.org/wiki/Software/systemd/logind

Thus, we need to return true here when the reply is neither 'no' nor
'na', in a similar fashion to what we do for the CanSuspend method.

[endlessm/eos-shell#4655]