Fix pressure barriers for hot corner on bottom
The existing code accounted for right vs. left hot corner
(taken from the right-to-left support from upstream)
but did not account for top vs. bottom hot corner
(which is not supported upstream at all).

Without this fix, instead of building pressure while pushing
against the bottom-right corner, the pressure would be continually
reset due to an incorrect determination that the corner was being left.

Previously, this case was not hit because the older version of X
did not support pressure barriers.  With the X upgrade, the pressure
barrier was implicitly enabled, and the broken direction check
prevented the hot corner from working anymore.

[endlessm/eos-shell#1325]