Add ./jasmine script to run tests
For ease of use, this adds a ./jasmine script that runs all the tests
against the uninstalled code, the same as "make check" does but with a
much more readable output.

Because of technical limitations with setting the mutter typelib dir and
requiring the "run-js-tests" loader, we can't achieve this by simply
executing the "jasmine" command like we do in other packages.

[endlessm/eos-sdk#3171]