Access static method on class, not instance
The old GJS allowed you to access static methods on an instance of the
class that defined them. Not so with the new GJS. So instead of
any_widget.get_default_direction(), for example, we now must do
Gtk.Widget.get_default_direction(). In this case the offending method
was Shell.WindowTracker.is_window_interesting().

[endlessm/eos-sdk#3265]