appDisplay: consolidate code for app icons activation
Differently from the desktop icon, the search icon code would not check
whether the application is already running before activating the splash
screen.
Consolidate this code in a separate function and call it from both
places.

This fixes a bug where the splash screen would appear for running
applications when launched from the search results.

[endlessm/eos-shell#5209]