Ignore speedwagon windows in the window-tracker
Make any speedwagon window fundamentally uninteresting. This means that
we'll wait until we see a non-speedwagon window before the app is
considered fully running.

Unfortunately, there are bugs in the startup-notification code which
mean that we won't see a titlebar icon, but that's not a major
regression from right now -- users can still see the splash page in the
overview and other places.