Unlock: do not map Esc key to Cancel
Since the keystroke that wakes the screensaver is applied to the
unlock dialog, if the user tries to use the Esc key to exit the
screensaver (which is one of the most intuitive keys to do so),
the dialog is immediately canceled by this single Esc keystroke.

In the future, we may want to block the wake-up keystroke from
being applied to the password, in which case we could revert
this commit and re-enable the Esc key to cancel the dialog.

[endlessm/eos-shell#5213]