appActivation: use a different special case for eos-link launchers
Previously we would always show a splash screen for weblinks, and
dismiss it after 700ms, regardless of whether the browser was running or
not. This does not match the behavior of any other launcher, which will
show the splash screen when the underlying application isn't running and
will not show it otherwise.

Change the special case logic we have for weblinks to implement the same
behavior for those launchers.

[endlessm/eos-shell#5547]