loginDialog: Always complete holds created in _showPrompt
If authentication is cancelled, the hold is "leaked," then when
a later authentication is successful, two holds will complete at the
same time, leading to strangeness as the second successful
authentication overrides the first. For example, during a password
reset, this causes the login screen to disappear immediately after the
user is prompted to confirm the new password, causing the new password
to be set before the confirmation and before being prompted to enter a
password hint (in the case where an authentication was previously
cancelled using the Cancel button under the login entry).

This is not an upstreamable fix because the login dialog has been
rewritten upstream to not use this arcane system of batches and holds
anymore.

[endlessm/eos-shell#4681]