util: Don't send duplicate reports of messages from the session worker
The login dialog understandably blows up if it receives two questions in
a row, as it expects to always send a response before it receives a new
question. Turns out we were reporting questions twice after an
authentication failure, because _retry() calls begin(), which
asynchronously triggers the creation of a new GdmUserVerifier to handle
the next authentication attempt. But the original GdmUserVerifier has
not been destroyed yet (though it might be eligible for garbage
collection at this point? not sure), and the ShellUserVerifier is still
listening to signals from it, so when the GdmSessionWorker sends an
InfoQuery or SecretInfoQuery, the ShellUserVerifier reports it once for
each GdmUserVerifier.

Fix this by explicitly destroying the original GdmUserVerifier. Make
sure not to clear the ShellUserVerifier, though, since it still needs to
display the authentication failed message.

[endlessm/eos-shell#4681]