loginDialog: Prevent horizontal size expansions
Currently the password hint and the prompt message both expand the
dialog significantly. Reduce the allocated to the password hint, so it
can't do this anymore, and ellipsize the prompt message to the same
width as the password hint.

Of course, the trade-off is that this results in vertical expansion
instead, but that doesn't look nearly so bad.

Note that this commit adds a missing style class that can in theory be
used by the login dialog if the session worker wills it (but I have
never seen it used before). This is just to stay on the safe side....

[endlessm/eos-shell#5895]