popupMenu: add scrollbars when content is too big
This patch is a general fix for too-tall popover menus
based on the last work with [endlessm/eos-shell#5926].
The problem with the previous approach is that it only
works with the date menu, while other menus still didn't
adapt to smaller screen sizes properly.

To fix that, add overlayed scrollbars on the PopupMenu
class directly. With that, every popup menu that goes
taller that the screen size is properly adapted. This
patch does not handle the case where screen is too narrow
for the menus, however.

[endlessm/eos-shell#5927]