power: fix labels alignment
Following the previous work porting the popup menu
to fit small screen sizes, the end-aligned labels
in the power menu were incorrectly aligned. In fact,
they were not expanding at all.

Fix that by putting the labels in a expanding StBin,
which correctly expands and aligns the label at the
end.

[endlessm/eos-shell#5927]