appDisplay: Fix app folder close button's position
The app folder's close button is positioned on the top right of the
folder's boxPointer. However, when the boxPointer has its arrow on
top (when the folder pops downwards), its box will be positioned lower
and thus the close button appears a bit higher than it should be.

These changes fix that problem by placing the close button taking
into account the position of the boxPointer.

[endlessm/eos-shell#6263]