popupMenu: adjust size and show the relevant scrollbars when needed
At the moment, these popup menus only show vertical scrollbars when
needed but never horizontal ones, which can't be simply set to use
and AUTOMATIC policy as that won't get the desired effect.

This patch adds some extra code to manually compare the preferred
size for the popup, compare it against the available space and decide
when to show the scrollbars and adjust the size of the popup, in
separate ways for each orientation.

[endlessm/eos-shell#6308]