popupMenu: adjust padding for horizontal scrollbars
At the moment, horizontal scrollbars are rendered far away from the
bottom edge of the popup widget, unnecessarily obscuring the widgets
that they are suppose to make easier to see.

This patch adds a new style to be applied to the internal box (from
PopupMenuBase) so that we can move the 1em vertical padding there
and out of the scrolled view, so that the horizontal scrollbars are
rendered as close to the edget as the vertical ones.

[endlessm/eos-shell#6308]