appDisplay: Update icons when their app's info changes
These changes connect the AllView to the new app-system's
"app-info-changed" in order to reload an app's icon.
This fixes the issue of adding a website to the grid, which can be added
with a generic icon until the real favicon is retrieved and set in the
respective desktop file, in which case the icon in the grid needs to be
reloaded.

[endlessm/eos-shell#4920]