appIconBar: Ensure pinned apps do not move around
Currently apps get pinned to the back of the favorites list, so that
when the shell is restarted they appear left-to-right in the order of
which apps were most recently pinned.

That's not the desired user experience. Once pinned, an icon now
remains on the taskbar, in the same position relative to other pinned
icons, even after rebooting.

[endlessm/eos-shell#6319]