shell-app: Fix setting the "starting" state
Currently only a few apps go into "starting" state, triggered by the
"startup sequence". This patch attemps to set the app state to
"starting" when the app is launched while in "stopped" mode.
It also allows the shell_app_sync_running_state to go into "running"
mode when the current mode is "starting".

[endlessm/eos-shell#4190]