workspaceMonitor: remove tracking of visible apps
Instead of explicitly tracking the visible apps, calculate them on the
fly when _updateOverview() is called.
While on one hand this may seem wasteful, tracking visible applications
in the workspace monitor is quite tricky, since the state needs to be
properly initialized e.g. when the shell process is restarted.
Furthermore, ShellAppSystem already does the heavy-lifting for us by
keeping track of the running applicaitons, so we never risk an unbounded
loop.

[endlessm/eos-shell#4190]