layoutManager: toggle Overview when text scale factor changes
The Shell's Overview sets up a clone of the app grid to apply
the Saturation effect when minimized. The problem with this
approach, however, is that we end up never updating the real
child when something changes while we're not in Overview.

This wouldn't be a problem, but there may be unmaximized windows
that exposes part of the Overview - in this case, this issue
is pretty noticeable.

Fix that by toggling the Overview on and off.

[endlessm/eos-shell#5950]