appActivation: Prevent showing windows before speedwagon
When the overview is hidden, the actors corresponding to visible
(non-minimized) windows are shown. As a result, the speedwagon's
animation does show up as smooth as it should.

These changes hide the mentioned actors while the speedwagon is shown
and show them again just before the speedwagon disappears. This way the
speedwagon's animation is smooth and there is no impact on the behavior
after it is gone.

https://phabricator.endlessm.com/T5419