AppIconMenu: Do not include speedwagon windows
Following the changes that prevent the windows menu from being activated
because of the presence of the speedwagon, it should also be explicitely
not included in the menu itself. Basically this prevents the menu from
showing the speedwagon at all, even if the menu is only visible if there
is more than one regular app window (not counting the speedwagon).

https://phabricator.endlessm.com/T6408