screenShield: Various fixes in _onUserBecameActive
Simplify the implementation. Most of this complexity is not needed.

Remove incorrect checks that were added to prevent the lock screen from
improperly showing when automatic lock is disabled. This code should
instead just check if the screen is actually locked or not.

We need to cancel the lock timeout when the user becomes active before
the screen locks, so it does not unexpectedly lock later on.

https://phabricator.endlessm.com/T6221