screenShield: Do not activate after showing the lightbox
In upstream GNOME, the screenShield is split into two stages: active
(screen shield is shown, may or may not be locked) and locked. If the
shield is locked, then after the screen shield is lifted, the lock
screen will be visible; otherwise, the shield lifts directly to the
desktop without requiring unlock.

In Endless, there is no separate screen shield: it is the lock screen.
So, whereas the upstream code should always activate the screen shield
after showing the lightbox, we do not want to do so, as that breaks the
Privacy panel setting that allows the user to specify a set amount of
time to wait before locking the screen. So... just don't do it. In this
case, it's desired and expected that the user return directly to the
desktop.
https://phabricator.endlessm.com/T6221