workspace: properly check for unmapped actors
When dealing with unmapped actors, Shell crashes since
St cannot create a theme node and ends up crashing. It
affects important applications such as eos-tutorial.

The proposed solution introduced another issue when an
unmapped window shows up since it completely aborts the
positioning code, which is something we want to do in the
majority of cases.

This commit places the check for unmapped windows before
adding the window, which fixes both the crash and the hot
corner issue.

https://phabricator.endlessm.com/T11146