viewSelector: redraw the clones on large text setting changes
This commit is a workaround on the Clutter misbehavior with
Framebuffer Objects. Precisely, the clone won't update properly
because the original actor mistakenly doesn't detect the living
clone and doesn't update correctly.

Fix that by adding a redraw'ing routine in an idle, to give
time for the original actor to update.

https://phabricator.endlessm.com/T5950