allView: consistently calculate the all view height
The current allocation code queries the allocation of
two different actors (the icon grid and the all view
itself) to calculate the sizing of the apps grid page
layout.

The problem with this approach is when the icon grid
outputs a size different than the all view actor, leading
to a miscalculation and making the search entry be
positioned wrongly.

Fix that by only calculating the size of the all view
actor.

https://phabricator.endlessm.com/T6047