main: initialize layout manager before workspace monitor
WorkspaceMonitor connects to a bunch of signals that depends
on the window manager. These signals are not synchronized in
any way with the rest of the codebase.

Given a fast-enough initialization, the WorkspaceMonitor instance
can try to load the overview before the overview is ready, since
both the overview and the layout manager are initialized after
WorkspaceMonitor.

Fix that by initializing the layoutManager and overview before
the WorkspaceMonitor, ensuring that we always have a valid overview
to animate.

https://phabricator.endlessm.com/T12522