missionGame: Communicate with the service to get the intro lesson
This lesson displays in the chatbox. We add a signal handler to
open-state-changed, which calls _showTaskDescriptionForLesson.
Every time the user puts something into the chatbox, it is passed
through "evaluate" and the result may be passed back to the service
if there is still an active lesson.

We take the result of that lesson and use it to determine whether
to show the "success" or "fail" message. If the "success" message
is to be shown, then the lesson counter is incremented and we
request the next task description.

Once the lesson counter has been exhausted, we display
the "done" message that was saved in this._introLesson and reset
the state back to its intial version.

Note that this is ripe for refactoring, which will likely
come later.