missionGame: Fix line-break algorithm in cases where \n was encountered
Previously when we encountered a '\n' character we treated it the
same as if we had encountered a word that was too long - the word
before the \n was being put on to the next line.

This behaviour was incorrect - the line is meant to break at the
\n, not at the word before it.

As a result of fixing this, we need to remove all newline
characters from appended text, since we're handling it internally
by splitting the line.