Default to pinning app center rather than browser
We want to ship with one app pinned as an example to help
users realize that the pinning functionality exists.

The app center is a better candidate for this than the browser,
as it is relevant to all users, including those who are completely
offline, and we want to help draw users into exploring the app center.

https://phabricator.endlessm.com/T13086