appIconBar: change allocation scheme for icons scrolling
Instead of updating the scrolled icons offsets both when icons are
scrolled and during the allocation cycle, just queue a relayout when
that happens, since whether icons are scrolled or not will affect the
allocation of the container.
This allows us to remove an extraneous iconSize in the size requisition
cycle of the content.

https://phabricator.endlessm.com/T13086