global: Add shell_global_window_matches_xid
Some discussion is needed here as to whether this is a good API to
have. It is convenient in the sense that it allows us to send
XIDs over D-Bus and match windows that way, but also could go against
GNOME-Shell's overall design because it seems quite clear that
the shell JS should not be allow to figure out what a window's
XID is and then punch through the abstraction.

I think shell_global_window_matches_xid is a good middle ground -
it allows you to check if a window matches an ID passed to you
externally, but doesn't allow the JS to figure out what the
XID is otherwise.